+++ src/main/php/net/stubbles/webapp/OptionsRoute.php
@@ -65,7 +65,7 @@ class OptionsRoute extends AbstractProcessableRoute
      *
      * @return  bool
      */
-    public function requiresRole()
+    public function requiresAuth()
     {
         return false;
     }
@@ -73,11 +73,23 @@ class OptionsRoute extends AbstractProcessableRoute
     /**
      * checks whether this is an authorized request to this route
      *
+     * @param   AuthHandler  $authHandler
      * @return  bool
      */
-    public function getRequiredRole()
+    public function isAuthorized(AuthHandler $authHandler)
     {
-        return null;
+        return true;
+    }
+
+    /**
+     * checks whether route required login
+     *
+     * @param   AuthHandler  $authHandler
+     * @return  bool
+     */
+    public function requiresLogin(AuthHandler $authHandler)
+    {
+        return false;
     }
 
     /**